Renamed a file.
